feat: throw on send failure (#3630)
On the IC, the act of making a call to a canister function can fail, so that the call cannot (and will not be) performed.
 This can happen due to a lack of canister resources, typically because the local message queue for the destination canister is full,  or because performing the call would reduce the current cycle balance of the calling canister to a level below its freezing threshold.
Such call failures, instead of trapping, are now reported by throwing an `Error` with new `ErrorCode` `#call_error { err_code = n }`, where `n` is the non-zero `err_code` value returned by the IC.
Like other errors, call errors can be caught and handled using `try ... catch ...` expressions, if desired.

 The constructs that now may throw call errors, instead of trapping as with previous versions of Motoko, are:
    * calls to `shared` functions (including oneway functions that return `()`).
      These involve sending a message to another canister, and can fail when the queue for the destination canister is full.
    * calls to local functions with return type `async`. These involve sending a message to self, and can fail when the local queue for sends to self is full.
    * `async` expressions. These involve sending a message to self, and can fail when the local queue for sends to self is full.
    * `await` expressions. These can fail on awaiting an already completed future, which requires sending a message to self to suspend and commit state.

(On the other hand, `async*` (being delayed) cannot throw, and evaluating `await*` will at most propagate an error from its argument but not, in itself, throw.)

# Implementation

Add explicit checks for message send failures, reified as catchable errors (exceptions).

* saves status of `ic0.call_perform` to new (internal) global
* test global after each message send, of which there are several flavours (async expressions, message sends and dynamic calls (a.k.a. `call_raw`), awaits on already completed promises)

Most of the transformation happen in ir_passes/await.ml where we have access to the failure/throw continuation, if any.


# Tasks

- [x] message sends (request and oneways)
- [x] async expressions 
- [x] raw calls (dynamic calls)
- [x] implicit yield on completed future (yuck).  Note this means not every await will commit, since, in (not so) rare cases, the self send queue will be full when trying to artificially suspend on a completed future. Not sure what to do in that case, trap?
- [x] test local asyncs (might be broken)
- [x] decide on error message and/or introduction of a new error code (would be a breaking change).
- [x] figure out how to fix the effect system (async can now throw synchronously, so need to have effect 'Await' to always be in cps context where throws can be handled)
- [x] update documentation for Error.mo in motoko-base.
- [x] split monolithic test into separate tests for async, send, await etc.

# Open Issues (Feedback wanted)

One minor worry with  this PR is that it is a breaking change. Previously a block of code that sends multiple messages, one of which fails to send, would trap, rolling back all the sends to the previous commit point. Now, the successful sends will remain enqueued, and the error/exception will flow to the nearest handler. If there is none, the method will exit with a canister_reject error, which will still send the other enqueued messages and commit state changes.


This is a breaking change, so we should release with 0.8.0.

I've added an opt-out compiler flag `--throw-on-call-error` to restore the old behaviour for users who don't care to audit their code.