bugfix: fix for quadratic import of actor classes (#3758)
The introduction of system class constructors had the unfortunate side-effect of making actor class libraries impure (due to a partial application at top-level of the 'system' class constructor to define the default class constructor). This meant that the compiler was no longer dead-coding them leading to quadratic behaviour in code generation: each actor class would include the code of all preceding actor classes, even if not actually used.

This PR make the translation of actor classes pure again, by defining the default constructor as a delayed application of the system constructor, allowing unused actor classes to be dead-code eliminated by the back-end.

In particular, reduces the code size of the large actor on the forum with six imported actor classes from 11MB to 3MB, 2.2MB after wasm shrink and 600K after gzip.

Fixes the bug reported here:
https://forum.dfinity.org/t/motoko-wasm-repeated-candid-declarations/18306

The original regression was introduced here:
https://github.com/dfinity/motoko/pull/3386

 * [x] the delay introduces an additional async - lets remove that using async*/await* and a readable helper in `internals.mo`
 * [x] the interpreted code still needs adjusting (see FIXME).

Closes #3769