chore: adopt `nixpkgs` 22.11 (#3600)
This is mostly mechanical, though it appears that `js_of_ocaml` v4.1.0 has a bug (when miscompiling `moc.js`), so we downgrade it to the known-good v4.0.0.

This needs to be followed-up by some differential debugging to figure out why v4.1.0 fails, and then potentially file bug report(s).

Another wrinkle is that the `musl`-version (static) if `python.twisted` doesn't check on Linux, so it has been worked around by disabling checking on (OCaml) `atdgen`.