perf: allocator fast path (#3090)
This introduces a XOR trick to detect whether the recent allocation stays on the same (already) allocated page, and only runs the expensive `grow_memory` route if a page boundary is crossed. It also reserves the last page as unallocated for essentials (like a potential future emergency allocation of the stable variable streaming buffer). The unallocated page can also serve as a guard page against stack overflow of the C/Rust RTS stack.

To make the trick work, a slight change in page allocation is necessary. When the new `HP` points to exactly the start of the page (i.e. has form `0xNNNN0000` for some `NNNN`) we have to eagerly allocate the page corresponding to all `0xNNNNxxxx`, because for subsequent in-page allocations the fast-path is in effect, and assume that the page is already allocated. Notably this rule _doesn't_ extend to new `HP = 0xFFFF0000`, as the last page is taboo.

### TODO
- [ ] make a test that successfully allocates till `0xFFFF0000` — dubious if this can be accomplished without a new primitive
- [ ] make a test that fails allocating till `0xFFFF0004`
- [x] consider `/ WASM_PAGE_SIZE` instead of `>> 16`
- [ ] try to increase cycles to make `oom-query.mo` work again