Backend: Avoid pointless bounds checking array access (#3853)
this has been in the back of my mind for a while: Originally, I intentionally used safe array access almost everywhere in the backend, even in places where the bounds should always be fine, so that bugs can be reported more efficiently. It seems that people have not found bugs related to that code, so we can drop the bounds check in compiler-generated loops over arrays (mostly in serialization/deserialization).

Turns out all array accesses that need to be bounds-checked go via the bignum index function anyways, so I moved the bounds check there.