perf: peephole-optimise `const 1; and; if t e` to `ctz; if e t` (#3873)
Masking the LSBit happens often when checking for _vanilla_ pointers, so this is probably a profitable optimisation.
_Indeed_: this improves cycle counts about 1% on our test set.

Instead of masking with 1, we count the trailing zeroes. We expect a subsequent `if`, but its legs must be switched, since a masked 1 corresponds to _no_ trailing zeroes.

I expect bignum operations and `Nat/Int/32/64` operations to be significantly cheaper now.

Kento checked that `binaryen`'s `wasm-opt` doesn't perform this substitution currently, but I think it could be contributed.