bugfix: check `@nextExpiration` after the changes are committed (#3883)
There was a slightly buggy behaviour with setting the global timer from `@timer_helper`. We observed that for each one-off timer set there were two expirations happening. The reason is that we were computing the next expiration before the current batch of expired timers were expunged. This re-added the same time as before, and thus caused the second (in-vain) wakeup.

The fix is of course to compute the next global expiration _after_ the already expired nodes were expunged.