bugfix: fix for #3910 (bad compilation of local async functions that pattern match on arguments) (#3916)
Fixes a regression introduced by the `async*/await*` feature and changes to the await translation.

The bug was trigged by a pattern matching local async function: `desugar.ml` pushes the pattern match into the body of the function, but outside the async expression, obscuring the async body from await.ml.

The fix is to spot that pattern in the `await.ml` transformation.

This fix is better than #3911 because it doesn't change the semantics of pattern match failure for local functions as #3911 does.
#3911 would delay the pattern match to the body of the `async/async*` where it would be reflected as an error for `async`, or merely delayed for `async*`, neither of which matches the eager trapping one would expect from a local call and the old, pre-`async*/await*`, semantics.

(This isn't an issue for shared `async` functions because we require their patterns to be irrefutable, so we can push the pattern match into the remote body without consider traps at all).

I've minimized the test case reported in #3910 as new regression test `run-drun/issue-3910.mo`