perf: improve codegen for `?_` and `#variant _` (and also `()`) (#3957)
Applies a shortcut when the inner pattern is a `WildP`. I also cover `()`, the prototypical non-binding irrefutable pattern.

Generally we now don't even make an attempt to enter into matching irrefutable patterns below options or tags when there are no bindings to fill.

This removes the unnecessary locals holding the scrutinees, since we don't have to project out for bindings or further pattern matching.

Normally the _push/drop_ peephole optimisation would _almost_ clean this up, but the local `tee` still remains.

Spotted in #3955.

Further optimisation could be to also allow:
- [x] _irrefutable_ patterns with `Freevars.(M.is_empty (pat p))` (e.g. `#foo ((), _)`) — see c06a0c9af
- when the scrutinee is _singleton-typed_ with no bindings in the pattern (but that might be less interesting).