feat: primitive for `ic0.canister_version` (#4027)
This PR adds support for the canister versioning feature of the IC, establishing the `canisterVersion` primitive.
A follow-up PR will amend the `create_canister` and `install_code` management canister calls (which we use to manage actor classes) with this information.
Retrieving the version history of canisters will be implemented from the `motoko-base` library.

- [x] tests added