temporary: no `or`-pattern inference (#4012)
As #3993 noted, our algorithm for inferring types of `or`-patterns is flawed, and its result is rejected by the IR type checker. Since IR-checking can be disabled, unsoundness may creep into released binaries (see #4001).

The conservative solution for this is to disallow pattern type inference for function definitions and require argument type annotations for now. This will be reverted as soon as we have a sound inference algorithm in place. (It is possible, as OCaml demonstrates with polymorphic variants).

There are attempts to mitigate the situation, e.g. with #4026.