chore: perform scheduled builds of `drun` in a sparse artefact environment (#4043)
Recently imports of `drun` tend to fail building, because the storage is very constrained on Linux builders, and `nix` preloads all necessary binary artefacts for the whole build process. This meant that when the `drun` build started there were already `OCaml` and `ghc` as well as many instances of `llvm`, `clang` and `rustc` present.

This PR detects when `drun` is about to be built, does it in a frugal way, and also cleans up after itself.

If this method turns out to be effective in the middle term, we can adopt it for `rustc` and other bigger chunks too.