Incremental GC: Optimize occupied size computation (#4021)
This is a minor optimization for the incremental GC:
Currently, the occupied heap size is computed by iterating through the partition table which is performed on each GC scheduling point. This optimization caches the heap size, by precomputing the occupied size in all partitions except for the dynamic space of the current allocation partition (where the dump allocator is active).

Instructions executed on GC scheduling:
Unoptimized | Optimized | Saving
-------------| --------------- |-----
5899 | 322 | 95%

In the GC benchmark, the performance gain is, however, negligible. It could be more relevant when moving to 64-bit.