chore: change binary release artefact names (#3997)
Here is a suggested PR to update the name of the release so they follow `$(uname -s)-$(uname -m)` convention. This would make it easier to fetch the releases from container build files like `Dockerfile`.

Hence: 
-  x86 Linux becomes `Linux-x86_64`
-  x86 Mac becomes `Darwin-x86_64`

In the future this would allow for: 
- M1/M2 Mac to become `Darwin-arm64`
- arm64 Linux to become `Linux-arm64`

Which would not have been possible under the current naming scheme which does not differentiate for the different macOS architecture (understandable as from my understanding there are no releases for M1/2 macOS yet).

_Note_: the legacy naming is still available. This will be discontinued when all downstream consumers have updated their procedures.