perf: keep a skewed pointer in `HP` (#4095)
Keeping the `HP` as _always skewed_ allows us to eliminate the subtraction on each individual allocation, at the expense of (un)skewing in a few seldom operations (which possibly can also be transitioned to skewed pointers sometime). Removing an instruction from an already compact _hot path_ has profound effects, as already has been demonstrated by #4064. This PR realises one of the suggestions from there.

The runtime savings amount to 1.5% on allocation-heavy benchmarks, and 0.3% on average, which is rather nice.

_Note_: Also included are a few `u32` -> `usize` changes, which may come handy in the 64-bit world.