feat: allow controllers to call `__motoko_stable_var_info` query endpoint (#4103)
This was originally spec-ed, but there was no possibility to implement it (a canister couldn't get its own controller list).
In the meantime canisters can get their controller list (via query), but a query cannot call another query 🫤.

So we had to wait until the `ic0.is_controller` syscall became available. Now it is there and the implementation is easy.

(It was technically possible to call it before, but it invariably trapped.)