chore: pull in the `ic-wasm` tool (#4109)
In `$PERF` mode `run.sh` now checks for an available `ic-wasm` and additionally runs the tests with `shrink -O3` optimisation. `nix-shell` now also builds this tool.

An attempt was made to include `ic-wasm` into the _`niv` updater_ and _hash update_ mechanism, so we shouldn't suffer the manual bumping/hash-correction overhead. It remains to be seen if this really triggers and works...