perf: extend the overflow detection mask to the right if the bump increment is a power of 2 (#4096)
Trying out the idea in https://github.com/dfinity/motoko/pull/4064/files/4eb3dfeb7fdc10737605c631edff33b3f0976eff#r1250740318.

TLDR: The page-crossing detection mask can be extended for power-of-2 increments by one bit (to the right) which allows for accurate detection of not only `HP` remaining on-page, but also page boundary crossings. This enables the property that the slow path gets entered only when necessary.

~~Alas, it is flawed... with a `HP` that can point exactly to the boundary between pages.~~

With #4095 merged this cannot happen. Let's prove it with pen and paper...

Assume the allocation increment is `N` = 8, the smallest possible (byte-)increment that is a power of 2 when allocating objects. In hex `N = 0x0008`, thus the unextended overflow mask is `M = 0xFFF0`, and we assume that such a mask guarantees (by #4095) the implication `(HP + N) & M > 0 => "i8.store (HP + N) x" won't crash` for all `HP` of the form `0bXX..X11`. In this PR we have to strengthen the implication to the extended mask `M' = M | N`, i.e. one bit to the right. We proceed as follows:

We only need to consider the case where `(HP + N) & M = 0`, otherwise we are done by the above assumption. This reduces
- `(HP + N) & M' > 0 => "i8.store (HP + N) x" won't crash`
-  via `(HP + N) & (M | N) > 0 => "i8.store (HP + N) x" won't crash`
- to `(HP + N) & (M | N) = N => "i8.store (HP + N) x" won't crash`

so the bit under the submask `N` must be set and under `M` no bit is set. This gives us a good picture of `HP + N`, namely `0bPPP..P0000000000001X11` where the 16 `P` bits are irrelevant and `X` can be 0 or 1. This also means that `HP` must have been `N` less: `0bPPP..P0000000000000X11`. Since `HP` was the result of a previous increment, and thus `"i8.store HP x" doesn't crash`, writing a byte to `0bPPP..P0000000000001X11` won't crash either (since it is on the same page). This finishes the implication.

Choosing a different `N` — that is a power of 2 and `< 0x8000` — will result in a similar proof, only that there are more `X` bits.