Incremental GC: Adjust Total Allocation Statistics (#4160)
# Incremental GC: Adjust Total Allocation Statistics

This is a minor, mostly cosmetic fix for the incremental GC: The total allocation statistics (`rts_total_allocation()`) was not monotonously growing but temporarily growing and shrinking while a GC run involved multiple increments. The reason was that the evacuation space was incorrectly counted as new allocations.

This PR does not imply any significant performance impact according to the GC benchmark measurements.

Before this PR (example of `cancan` GC benchmark case):

![image](https://github.com/dfinity/motoko/assets/112870813/35d0f8f4-e909-4a1b-86e7-2757468f3b56)

**Fix**: The temporary copying space for evacuations is no longer considered in the total allocation metric.

With this PR (example of `cancan` GC benchmark case):

![image](https://github.com/dfinity/motoko/assets/112870813/4808b517-c63b-4761-b168-45a5262b403d)