bugfix: fix candid decoding at immutable array types to support opt defaulting (#4240)
Another Candid decoding bug, seems independent of #4238. 

I think the code for decoding immutable arrays was never extended to handle backtracking correctly.
When recovery is an option (haha) and the type descriptor does not match, the current code would continue to try to decode the array length and elements. When recovery is not an option, the previous trap would render this code unreachable. But when recovery is enabled, we need to bail and not try to read the array size and elements but just return the sentinel value instead.

- [ ] Consider applying the same refactoring to mutable array, if possible. 
      Since these aren't Candid, but extended Candid (for stable variables), and don't need to support recovery, the change is (probably) not necessary but nice for uniformity only.
- [ ] restore or delete original repro - I must have edited it while experimenting