bug: fix and repro for issue-4236 (bad renaming of `or`-patterns) (#4244)
Fix `ir_def/rename.ml`, for renaming bound variables, to properly handle our newly extended `or`-patterns.

With `or`-patterns, the same variable may have several binding occurrences, so renaming has to treat all of these consistently.
Instead of renaming variables on the fly, for `or`-patterns, we now determine the set of bound identifiers bound by all the 
patterns beneath the `or`, rename them, and substitute consistently within the `or`-pattern to produce the renamed `or`-pattern.