bugfix: repro and fix for issue-4297 (#4306)
repro and (ugly) fix for issue #4297:

Moc let's you compile a sequence of declarations as an implicit actor, which is used by some of our drun tests and was stumbled on by a user. Unfortunately, we failed to set up the initial environment corrrectly, so those declarations could do async/await  at top-level (i.e. in the body of the actor), which should be illegal and rejected (while not being rejected for interpreted code, whose top-level is allowed to async/await).

The hacky fix is to inspect the progs being compiled using `CompUnit.comp_unit_of_prog` and set up the initial capability appropriately. Even better would be to pass the capability down somehow, but this turns out to be trickier than expected.

To avoid clobbering the existing tests too much, we also make typing functions that check the current static capability return local_errors (not errors) on failure, to continue with typechecking and report any other static errors.

Added bonus: setting up the capabilities properly prevents us from doing top-level async/await in the prelude, which would otherwise wreak all sorts of havoc in compile.ml due to unexpected renaming of well-known prelude functions due to cps conversion.