bugfix: fix bug affecting `--experimental-rtti` (#4434)
I introduced a bug when making rtti a compiler option.

In particular, in `compile.ml`, the value `BigNum.tag_padded` did not test the new compiler flag properly - it was missing the conventional `env` argument and produced code that was failing to tag compact `Int` and `Nat` values (doing a logical or with the non-rtti tag, i.e. 0).

It would probably be safer to read the flag from the env argument, not directly from `Flags`, but that can come in a follow-on PR.

The good news is that the bug is benign for vanilla moc (and we don't even advertise `--experimental-rtti`)

Ì told you it was experimental.