Exempt top-level identifiers from unused detection (#4452)
Exempt **top-level** program declarations from unused detection:
* Actor class identifiers
* Module identifiers
* Actor identifiers

This is mostly for developer convenience but also avoids unnecessary warnings in the IDE where files are separately checked, see below.

## Note
With Wasi mode, Motoko programs can contain other top-level declarations that are still reported as unused if so, e.g. unused top-level variables or functions. 

**Exception**: 
`let unusedModule = module {}` is no longer detected as unused if appearing as a top-level declaration.

## Fix
This also fixes https://github.com/dfinity/motoko/issues/4448: The unnecessary warning here appears because the files are separately checked in VSCode. When compiled as one program, e.g. with `dfx` using the original stricter detection, the actor class would not be reported unused as it is instantiated in the main program.