Split stdout and stderr in viper/* tests
The `moc --viper` command outputs generated Viper code to stdout and
diagnostic messages to stderr.

* Prior to this change, test/run.sh used the `run` helper to produce
  .vpr files. The `run` helper always merges stdout and stderr, which
  results in ill-formed .vpr files if there are any diagnostic messages.

* Now test/run.sh has an additional helper `run_stderr` that keeps
  stdout and stderr separate. By using it instead of `run` we ensure
  that .vpr files are not contaminated with diagnostics.

Run `make -C test/viper` to confirm that this fixes broken Viper tests.

Note that `run_stderr` is a copy of `run` with minor tweaks, so this
solution introduces some technical debt. A more principled approach
would be to separate stdout and stderr in all commands (i.e. wholly
replace `run` with `run_stderr`), but I opted not to do that at this
time to avoid a massive overhaul of the test suite.