[motoko-san] bugfix & remove static invs
Problem: we tried to deduce static invariants like size of immutable array
fields automatically and add them into actor's invariant. However we made a bug
and deduce it for mutable arrays as well and in the motivating use
case (reverse.mo) it could be not true indeed.

Solution: remove this deduction and add invariant on size of array in the
reverse.mo manually into actor's invariant as it should be.

Also see next commit about deduction invariants in loops