experiment: deprecate experimental stable memory (improved) (#4575)
Leverages `/// @deprecate` comments to deprecate uses of ExperimentalStableMemory.

Adds moc flag `--experimental-stable-memory <n>` to control access to legacy stable memory:

- n<0: error on use of (ESM) prims.
- n=0: warn on use of prims.
- n=1: warning-less use of prims.

Current default is n=0 to encourage users to migrate their code to use preferred `Region.mo` without preventing compilation.

The implementation adds /// @deprecate M0199 comments Prim.mo and spots this specific deprecation during type-checking, reporting this specific violation just once per compilation unit, as an error, warning or not at all, depending on compiler setting.

The net effect is that, by default, any code that directly or indirectly imports ESM will be reported, as will any code that directly access a stable memory primitive, alerting the user to potential vulnerabilities. The code will still compile, but in future we can prevent that by defaulting n to 0, not 1.

The not-so-nice aspect of this implementation is that:
1: The errors only indicate that ESM *may* be in use. 
2: They do not pinpoint which importing library is responsible for the warning, but adding `/// @deprecate M0199` to corresponding ESM fields directly should achieve that too.