IR: don't create unnecessary variable renaming for the sake of `NewObjE` (#4611)
`NewObjE` wants variable names for field values. When the expression is _already a variable_ then just reuse its name!

Also, if we managed to eliminate all necessary variable (re-)bindings, then the enclosing new scope (`BlockE`) is redundant.

By generating less IR we speed up compilation, and potentially create opportunities for further simplifications.

Note: To make this work I had to allow initialisation of `NewObjE` immutable fields from mutable variables. This necessitated changes in both the `interpret_ir.ml` and `check_ir.ml`. Now the interpreters and the backend behave the same.

TODO:
- [x] tests
- [ ] `var field = immutVar` (not possible — unless we intro a _varenv_ that teaches us that the source `immutVar` is a `LetD`, and even then the backend is tricky*)

If we don't want a varenv in `desugar.ml` then we must also admit `var field = mutVar` without the fresh `VarD` in the enclosing block, furthermore:
* Why is the backend tricky? `get_aliased_box` when called with an _rvalue_ would have to allocate a mutbox on the fly, fill it, and return the pointer.
* the IR-interpreter would have to examine the `{T.Mut/non-Mut}`⨯`{V.Mut/non-Mut}` matrix and do something like
``` OCaml
   let v = match f.note, Lib.Promise.value (find f.it.var env.vals) with
     | T.Mut _, V.Mut v -> V.Mut (ref !v) (* fresh mutable cell *)
     | T.Mut _, _ -> V.Mut reg
     | _, V.Mut v -> !v
     | _, v -> v in
```
`refD` is probably a problem, as it would allocate a fresh `V.Mut` with a reused `ref` in it, and then the first line would destroy the aliasing.

I don't want to think this through. (Examining the experimental flag (`= false`) would probably help.) But `var` fields are seldom, so just don't overcomplicate it.