optimisation: annotations to track (reads from) mutable declarations and adapt `Ir.VarE` (#4637)
By making reads from mutable variables distinguishable, we restore the referential transparency of `VarE (Const, id)`. This was the reason why #4611 failed (and had to be reverted in #4623).

This opens up a bunch of optimisation possibilities
- `VarE` path compression
- dead `LetD` elimination
- `BlockE`/`NewObjE` simplification
- nested `BlockE` merging (given the defs are disjoint and outer uses don't reference inner ones; use `freevars.ml`)
- etc.

In the long run this may allow us to extend the IR effect system with {`Load`, `Store`} effects, and thus pave the way for transactional memory.

We now have a note on `Syntax.VarE`'s identifier, being immutable by default. When performing type checking, this field can change to signify a mutable variable source after looking up the identifier in the environment. Then desugaring uses mutability info to create either `IR.VarE (Const, _)` (for immutable) or a`IR.VarE (Var, _) (for mutable). The IR type checker then asserts that the right kind of access is used.

The Wasm output should be unchanged.

_NOTE_: with this PR `git revert 13d9d5748700662d30d54d6ccfcf2787deac482e` can be done and it passes. So it seems to work 😀